backend/requirements.txt
fastapi[standard]
python-dotenv
google-genai
supabase
temporalio
yt-dlp